requirements.txt
agate==1.9.1
annotated-types==0.7.0
anyio==4.12.1
arcosparse==0.4.2
attrs==25.4.0
babel==2.18.0
boto3==1.42.66
botocore==1.42.66
certifi==2026.2.25
cffi==2.0.0
cftime==1.6.5
charset-normalizer==3.4.5
click==8.2.1
cloudpickle==3.1.2
colorama==0.4.6
copernicusmarine==2.3.0
cramjam==2.11.0
cryptography==46.0.5
daff==1.4.2
dask==2026.1.2
db-dtypes==1.5.0
dbt-adapters==1.22.8
dbt-bigquery==1.11.0
dbt-common==1.37.3
dbt-core==1.11.2
dbt-extractor==0.6.0
dbt-protos==1.0.440
dbt-semantic-interfaces==0.9.0
deepdiff==8.6.1
distro==1.9.0
docstring_parser==0.17.0
donfig==0.8.1.post1
fastjsonschema==2.21.2
fastparquet==2025.12.0
fsspec==2026.2.0
geopandas==1.1.3
google-api-core==2.30.0
google-auth==2.49.0
google-auth-oauthlib==1.3.0
google-cloud-aiplatform==1.141.0
google-cloud-bigquery==3.40.1
google-cloud-core==2.5.0
google-cloud-dataproc==5.25.0
google-cloud-resource-manager==1.16.0
google-cloud-storage==3.1.1
google-crc32c==1.8.0
google-genai==1.66.0
google-resumable-media==2.8.0
googleapis-common-protos==1.73.0
grpc-google-iam-v1==0.14.3
grpcio==1.78.0
grpcio-status==1.78.0
h11==0.16.0
h5netcdf==1.8.1
httpcore==1.0.9
httpx==0.28.1
idna==3.11
importlib_metadata==8.7.1
isodate==0.7.2
Jinja2==3.1.6
jmespath==1.1.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jupyter_core==5.9.1
leather==0.4.1
locket==1.0.0
MarkupSafe==3.0.3
mashumaro==3.14
more-itertools==10.8.0
msgpack==1.1.2
nbformat==5.10.4
netCDF4==1.7.4
networkx==3.6.1
numcodecs==0.16.5
numpy==2.4.3
oauthlib==3.3.1
orderly-set==5.5.0
packaging==26.0
pandas==2.3.3
pandas-gbq==0.34.0
parsedatetime==2.6
partd==1.4.2
pathspec==0.12.1
platformdirs==4.9.4
proto-plus==1.27.1
protobuf==6.33.5
psutil==7.2.2
pyarrow==23.0.1
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pydata-google-auth==1.9.1
pyogrio==0.12.1
pyproj==3.7.2
pystac==1.14.3
python-dateutil==2.9.0.post0
python-slugify==8.0.4
pytimeparse==1.1.8
pytz==2026.1.post1
PyYAML==6.0.3
referencing==0.37.0
requests==2.32.5
requests-oauthlib==2.0.0
rpds-py==0.30.0
rsa==4.9.1
s3transfer==0.16.0
semver==3.0.4
shapely==2.1.2
six==1.17.0
sniffio==1.3.1
snowplow-tracker==1.1.0
sqlparse==0.5.4
tenacity==9.1.4
text-unidecode==1.3
toolz==1.1.0
tqdm==4.67.3
traitlets==5.14.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
websockets==16.0
xarray==2026.2.0
zarr==3.1.5
zipp==3.23.0
